[ml] introduce `BaseEstimator` as another base class
