+++ src/pymor/algorithms/ml/vkoga/kernels.py
@@ -4,8 +4,10 @@
 
 import numpy as np
 
+from pymor.algorithms.ml.base_estimator import BaseEstimator
 
-class GaussianKernel:
+
+class GaussianKernel(BaseEstimator):
     r"""Standalone Gaussian (RBF) kernel with scikit-learn-like interface.
 
     .. math::
@@ -18,6 +20,8 @@ class GaussianKernel:
     length_scale
         The length scale parameter of the kernel.
     """
+
+    _params = ('length_scale',)
 
     def __init__(self, length_scale=1.0):
         assert length_scale > 0.
@@ -51,35 +55,3 @@ class GaussianKernel:
         """Return the diagonal of the kernel matrix."""
         X = np.atleast_2d(X)
         return np.ones(X.shape[0])
-
-    def get_params(self, deep=True):
-        """Returns a dict of the init-parameters of the kernel, together with their values.
-
-        The argument `deep=True` is required to match the scikit-learn interface.
-
-        Parameters
-        ----------
-        deep
-            Since the kernel has no subobjects with parameters, this parameter is only
-            required to match the scikit-learn interface via duck-typing.
-
-        Returns
-        -------
-        A dictionary of parameters and respective values of the kernel.
-        """
-        return {'length_scale': self.length_scale}
-
-    def set_params(self, **params):
-        """Set the parameters of the kernel.
-
-        Parameters
-        ----------
-        Kernel parameters to set.
-
-        Returns
-        -------
-        An instance of the kernel with the new parameters.
-        """
-        for key, value in params.items():
-            setattr(self, key, value)
-        return self
